content(es): merge duplicate monte fuji articles + redirect
- Merge transport/climbing content from monte-fuji-desde-tokio into
  monte-fuji-se-ve-desde-tokio (new sections: train, bus, climb vs view)
- Delete EsMonteFujiDesdeTokio.tsx
- Remove from AppRoutes and EsBlogIndex
- Add 301 redirect in netlify.toml
- Update cross-links in EsItinerarioTokio5Dias
- Remove from sitemap.xml
