+++ public/sitemap.xml
@@ -406,12 +406,6 @@
     <priority>0.6</priority>
   </url>
   <url>
-    <loc>https://tanuki-tabi-travel.com/es/blog/monte-fuji-desde-tokio</loc>
-    <lastmod>2026-03-13</lastmod>
-    <changefreq>monthly</changefreq>
-    <priority>0.6</priority>
-  </url>
-  <url>
     <loc>https://tanuki-tabi-travel.com/es/blog/propinas-en-japon</loc>
     <lastmod>2026-03-13</lastmod>
     <changefreq>monthly</changefreq>
