+++ src/AppRoutes.tsx
@@ -65,7 +65,7 @@ import EsMejorEpocaVisitarTokio from "./pages/es/blog/EsMejorEpocaVisitarTokio";
 import EsEtiquetaTemplos from "./pages/es/blog/EsEtiquetaTemplos";
 import EsValeLaPenaGuia from "./pages/es/blog/EsValeLaPenaGuia";
 import EsComparativaExcursiones from "./pages/es/blog/EsComparativaExcursiones";
-import EsMonteFujiDesdeTokio from "./pages/es/blog/EsMonteFujiDesdeTokio";
+
 import EsPropinasenJapon from "./pages/es/blog/EsPropinasenJapon";
 import EsQueSecomeEnJapon from "./pages/es/blog/EsQueSecomeEnJapon";
 import EsMercadoTsukijiTokio from "./pages/es/blog/EsMercadoTsukijiTokio";
@@ -156,7 +156,7 @@ const AppRoutes = () => (
         <Route path="/es/blog/etiqueta-templos-santuarios" element={<EsEtiquetaTemplos />} />
         <Route path="/es/blog/vale-la-pena-contratar-guia" element={<EsValeLaPenaGuia />} />
         <Route path="/es/blog/comparativa-excursiones" element={<EsComparativaExcursiones />} />
-        <Route path="/es/blog/monte-fuji-desde-tokio" element={<EsMonteFujiDesdeTokio />} />
+
         <Route path="/es/blog/propinas-en-japon" element={<EsPropinasenJapon />} />
         <Route path="/es/blog/que-se-come-en-japon" element={<EsQueSecomeEnJapon />} />
         <Route path="/es/blog/mercado-tsukiji-tokio" element={<EsMercadoTsukijiTokio />} />
