--- src/pages/es/blog/EsMonteFujiDesdeTokio.tsx
@@ -1,297 +1,0 @@
-// TRANSLATION REVIEW NEEDED: Please have a native Spanish speaker review this content before publishing
-import { Link } from "react-router-dom";
-import { ArrowLeft, Calendar, User } from "lucide-react";
-import { Layout } from "@/components/layout/Layout";
-import { SEO } from "@/components/SEO";
-import { RelatedTourCards } from "@/components/blog/RelatedTourCards";
-import { ArticleFooterCTAEs } from "@/components/blog/ArticleFooterCTAEs";
-
-const EsMonteFujiDesdeTokio = () => {
-  return (
-    <Layout>
-      <SEO
-        title="Monte Fuji desde Tokio: guía completa 2026 para verlo o subir"
-        description="¿Cómo ir al Monte Fuji desde Tokio? Un guía con licencia oficial explica las mejores opciones para verlo o subirlo, temporadas, precios y errores que evitar."
-        canonicalPath="/es/blog/monte-fuji-desde-tokio"
-        hreflang={[
-          { lang: "es", path: "/es/blog/monte-fuji-desde-tokio" },
-        ]}
-      />
-
-      {/* Hero Image */}
-      <section className="relative h-[40vh] md:h-[50vh] min-h-[300px]">
-        <img
-          src="/images/tours/hakone-ropeway-owakudani.webp"
-          alt="Vista del Monte Fuji desde Hakone"
-          className="w-full h-full object-cover"
-          loading="eager"
-          fetchpriority="high"
-          width={1600}
-          height={900}
-        />
-        <div className="absolute inset-0 bg-gradient-to-t from-black/60 via-black/20 to-transparent" />
-      </section>
-
-      {/* Article Header */}
-      <section className="pt-16 pb-12 bg-secondary/30">
-        <div className="container-section">
-          <div className="max-w-3xl">
-            <Link
-              to="/es/blog"
-              className="inline-flex items-center gap-2 text-sm text-muted-foreground hover:text-foreground transition-colors mb-6"
-            >
-              <ArrowLeft className="w-4 h-4" />
-              Volver al Blog
-            </Link>
-            <p className="text-label text-accent mb-3">Planifica tu Viaje</p>
-            <h1 className="heading-display text-foreground">
-              Monte Fuji desde Tokio: Guía Completa 2026 para Verlo o Subir
-            </h1>
-            <div className="mt-6 flex items-center gap-6 text-sm text-muted-foreground">
-              <span className="flex items-center gap-2">
-                <User className="w-4 h-4" />
-                Manabu, Guía con Licencia
-              </span>
-              <span className="flex items-center gap-2">
-                <Calendar className="w-4 h-4" />
-                7 de marzo de 2026
-              </span>
-            </div>
-          </div>
-        </div>
-      </section>
-
-      {/* Article Content */}
-      <section className="py-16">
-        <div className="container-section">
-          <article className="max-w-3xl mx-auto prose-custom">
-            {/* Introducción */}
-            <p className="text-lg text-muted-foreground leading-relaxed mb-4">
-              De cada diez viajeros que llegan a Tokio con la idea de ver el Monte Fuji desde Tokio, al menos seis se llevan una decepción. No porque sea imposible, sino porque nadie les explicó algo fundamental: el Fuji solo se deja ver cuando él quiere. En mis más de ocho años como guía con licencia, he aprendido que la diferencia entre una experiencia inolvidable y una frustración total está en la planificación, no en la suerte.
-            </p>
-            <p className="text-muted-foreground leading-relaxed mb-4">
-              Esta guía es exactamente lo que le cuento a cada cliente antes de organizar su excursión al Monte Fuji desde Tokio. Sin rodeos, sin frases bonitas de folleto turístico: datos reales, precios actualizados y los errores que veo repetirse semana tras semana. Si estás planeando tu viaje a Japón y el Fuji está en tu lista, esto te va a ahorrar tiempo, dinero y disgustos.
-            </p>
-
-            {/* ¿Se puede ver el Monte Fuji desde Tokio sin ir? */}
-            <h2 className="heading-section text-foreground mt-12 mb-6">
-              ¿Se puede ver el Monte Fuji desde Tokio sin ir?
-            </h2>
-            <p className="text-muted-foreground leading-relaxed mb-4">
-              Sí, se puede. Y a veces es espectacular. Pero no siempre. El Monte Fuji está a unos 100 kilómetros al suroeste de Tokio, y en días despejados su silueta nevada aparece detrás del skyline de la ciudad como un telón de fondo cinematográfico. El problema es que esos días despejados no son tan frecuentes como uno imagina.
-            </p>
-
-            <h3 className="text-xl font-medium text-foreground mt-8 mb-4">
-              Los mejores miradores dentro de Tokio
-            </h3>
-            <p className="text-muted-foreground leading-relaxed mb-4">
-              Si quieres intentar ver el Monte Fuji desde Tokio sin salir de la ciudad, estos son los lugares que yo recomiendo a mis clientes. La <strong className="text-foreground">plataforma de observación del edificio del Gobierno Metropolitano en Shinjuku</strong> es mi favorita: es gratuita, abre hasta las 21:30 y en un día claro ofrece una vista frontal del Fuji recortado contra el horizonte. Otro punto excelente es el <strong className="text-foreground">mirador de Shibuya Sky</strong>, en la azotea del Shibuya Scramble Square (entrada: 2.000 yenes para adultos). Desde ahí, el Fuji aparece al fondo en la dirección opuesta al famoso cruce. También funciona bien el <strong className="text-foreground">Tokyo Skytree</strong>, aunque la distancia hace que la montaña se vea más pequeña de lo esperado.
-            </p>
-
-            <h3 className="text-xl font-medium text-foreground mt-8 mb-4">
-              ¿Cuándo se ve mejor? (temporadas y condiciones)
-            </h3>
-            <p className="text-muted-foreground leading-relaxed mb-4">
-              La visibilidad depende de la humedad y la contaminación atmosférica. <strong className="text-foreground">Los mejores meses para ver el Fuji desde Tokio son de noviembre a febrero</strong>, cuando el aire frío y seco despeja el cielo. En verano, la humedad crea una bruma casi permanente que lo oculta incluso en días soleados. Las primeras horas de la mañana, justo después del amanecer, suelen ofrecer la mejor visibilidad antes de que la ciudad genere calor y vapor. Como guía, siempre les digo a mis clientes: si amaneció despejado en invierno, corran al mirador. Esa oportunidad puede no repetirse durante su estancia.
-            </p>
-
-            {/* Cómo llegar al Monte Fuji desde Tokio */}
-            <h2 className="heading-section text-foreground mt-12 mb-6">
-              Cómo llegar al Monte Fuji desde Tokio (opciones reales)
-            </h2>
-            <p className="text-muted-foreground leading-relaxed mb-4">
-              Si quieres verlo de cerca, tocarlo o subirlo, necesitas salir de Tokio. Hay tres formas principales de hacer la excursión al Monte Fuji desde Tokio, y cada una tiene sus ventajas según tu presupuesto, nivel de japonés y preferencias.
-            </p>
-
-            <h3 className="text-xl font-medium text-foreground mt-8 mb-4">
-              En tren (la opción más práctica)
-            </h3>
-            <p className="text-muted-foreground leading-relaxed mb-4">
-              Desde la estación de Shinjuku, el <strong className="text-foreground">tren expreso Fuji Excursion</strong> te lleva directamente a la estación de Kawaguchiko en aproximadamente 1 hora y 50 minutos, sin transbordos. El billete cuesta alrededor de 4.130 yenes por trayecto. Este tren tiene reserva de asiento obligatoria y se agota rápido, especialmente en fines de semana y temporada alta, así que te recomiendo reservar con antelación desde la web de JR East. Si tienes el Japan Rail Pass, puedes cubrir parte del trayecto (hasta Otsuki) y luego pagar solo el tramo de la Fujikyu Railway (1.170 yenes). Kawaguchiko es la base ideal para explorar la zona de los Cinco Lagos y disfrutar de las mejores vistas del Fuji reflejado en el agua.
-            </p>
-
-            <h3 className="text-xl font-medium text-foreground mt-8 mb-4">
-              En autobús (la más económica)
-            </h3>
-            <p className="text-muted-foreground leading-relaxed mb-4">
-              Los autobuses Highway Bus salen desde la terminal de Shinjuku Busta (sur de la estación de Shinjuku) y llegan a Kawaguchiko en unas 2 horas, dependiendo del tráfico. El precio es de unos 2.200 yenes por trayecto, casi la mitad que el tren. La desventaja es que el tráfico puede alargar el viaje considerablemente, sobre todo los fines de semana y festivos. En mis tours, yo suelo recomendar el autobús para ir y el tren para volver, combinando economía con comodidad.
-            </p>
-
-            <h3 className="text-xl font-medium text-foreground mt-8 mb-4">
-              Con tour organizado vs. por tu cuenta
-            </h3>
-            <p className="text-muted-foreground leading-relaxed mb-4">
-              Ir por tu cuenta es perfectamente viable si hablas algo de inglés y te manejas con Google Maps. Pero hay aspectos que solo un guía local puede resolver: saber a qué mirador ir según las condiciones del día, evitar las zonas masificadas, y combinar el Fuji con lugares cercanos que no aparecen en las guías turísticas estándar. En mis tours privados, suelo incluir paradas en Oshino Hakkai (un pueblo con manantiales de agua cristalina al pie del Fuji) o en Chureito Pagoda, la famosa pagoda roja con el Fuji de fondo, que pierde todo su encanto si llegas a la hora equivocada. La ventaja de un tour privado es que adaptamos el plan al clima real de ese día, algo imposible con tours grupales que siguen un itinerario fijo.
-            </p>
-
-            {/* ¿Cuándo es mejor ir? */}
-            <h2 className="heading-section text-foreground mt-12 mb-6">
-              ¿Cuándo es mejor ir? (temporadas y clima)
-            </h2>
-
-            <h3 className="text-xl font-medium text-foreground mt-8 mb-4">
-              Temporada de escalada (julio-septiembre)
-            </h3>
-            <p className="text-muted-foreground leading-relaxed mb-4">
-              El Monte Fuji solo se puede escalar oficialmente entre principios de julio y mediados de septiembre. Fuera de ese período, las rutas están cerradas por seguridad: hay nieve, hielo y condiciones meteorológicas peligrosas. La ruta más popular es la <strong className="text-foreground">Yoshida Trail</strong>, que parte de la Quinta Estación (2.300 metros de altitud) y llega a la cumbre a 3.776 metros. La subida toma entre 5 y 7 horas, y la bajada entre 3 y 4 horas. Desde 2024, existe un sistema de reservas obligatorio y una tasa de acceso de 2.000 yenes por persona para controlar la masificación. Las plazas se agotan, así que si planeas subir, reserva con semanas de antelación.
-            </p>
-
-            <h3 className="text-xl font-medium text-foreground mt-8 mb-4">
-              Fuera de temporada: el Fuji sin subir
-            </h3>
-            <p className="text-muted-foreground leading-relaxed mb-4">
-              Paradójicamente, las mejores experiencias visuales del Monte Fuji desde Tokio y sus alrededores ocurren fuera de la temporada de escalada. <strong className="text-foreground">De noviembre a marzo</strong>, la montaña luce su capa de nieve completa, el aire es más limpio y las multitudes desaparecen. El otoño (noviembre) regala la combinación del Fuji nevado con los colores rojos y dorados de los arces. La primavera (finales de marzo a mediados de abril) ofrece la postal más icónica del mundo: el Fuji nevado con cerezos en flor en primer plano. En mis tours de temporada baja, a menudo tenemos los miradores prácticamente para nosotros solos.
-            </p>
-
-            {/* Subir vs. solo verlo */}
-            <h2 className="heading-section text-foreground mt-12 mb-6">
-              Subir vs. solo verlo: ¿qué te recomienda un guía?
-            </h2>
-            <p className="text-muted-foreground leading-relaxed mb-4">
-              Esta es la pregunta que más me hacen. Y mi respuesta honesta es: depende de para qué vienes a Japón. Subir el Fuji es una experiencia física exigente que consume un día entero (o dos, si duermes en un refugio de montaña). Si tu viaje es de 7 a 10 días y quieres conocer Tokio, Kioto y Osaka, dedicar un día completo a la subida puede significar sacrificar experiencias culturales que no volverás a tener.
-            </p>
-
-            <h3 className="text-xl font-medium text-foreground mt-8 mb-4">
-              Si decides subir: lo que necesitas saber
-            </h3>
-            <p className="text-muted-foreground leading-relaxed mb-4">
-              La escalada no requiere experiencia técnica, pero sí preparación física. El sendero es empinado, con tramos de roca volcánica suelta, y la altitud afecta incluso a personas en buena forma. Lo esencial: lleva ropa de abrigo para la cumbre (puede haber 0 grados incluso en agosto), un frontal para la caminata nocturna si quieres ver el amanecer, al menos 2 litros de agua por persona, y comida energética. Los refugios de montaña (yamagoya) ofrecen cena y un espacio para dormir por unas 8.000-10.000 yenes, pero son espartanos: dormitorios compartidos de tatami donde duermes hombro con hombro con desconocidos. Reserva con meses de antelación, especialmente para los refugios de la Ruta Yoshida.
-            </p>
-
-            <h3 className="text-xl font-medium text-foreground mt-8 mb-4">
-              Si decides solo verlo: las mejores experiencias
-            </h3>
-            <p className="text-muted-foreground leading-relaxed mb-4">
-              Ver el Monte Fuji desde Tokio o desde la zona de los Cinco Lagos sin subirlo puede ser igual de memorable, o más. Mis experiencias favoritas para recomendar: un paseo en barco por el <strong className="text-foreground">Lago Kawaguchiko</strong> con el Fuji reflejado en el agua; la subida (corta y fácil) al <strong className="text-foreground">mirador de Chureito Pagoda</strong> en Fujiyoshida, con sus 398 escalones y la vista más fotografiada de Japón; y un baño en un onsen (aguas termales) con vista al Fuji, como el Fuji Yurari Onsen, donde te relajas en agua caliente al aire libre mirando la montaña. Esta última experiencia, como les digo a mis clientes, es probablemente lo más japonés que puedes hacer. Imagina algo así como ir a un balneario, pero con la montaña más sagrada de Japón como paisaje. Si incluyes esta zona en tu{" "}
-              <Link to="/es/blog/itinerario-tokio-5-dias" className="text-accent hover:underline">
-                itinerario de 5 días en Tokio
-              </Link>
-              , un día completo en la zona del Fuji encaja perfectamente.
-            </p>
-
-            {/* Lo que nadie te dice */}
-            <h2 className="heading-section text-foreground mt-12 mb-6">
-              Lo que nadie te dice antes de ir
-            </h2>
-
-            <h3 className="text-xl font-medium text-foreground mt-8 mb-4">
-              Errores comunes que veo en mis clientes
-            </h3>
-            <p className="text-muted-foreground leading-relaxed mb-4">
-              Después de cientos de tours, estos son los errores que se repiten una y otra vez. <strong className="text-foreground">Primer error: reservar la excursión al Monte Fuji desde Tokio con un solo día de margen.</strong> Si llueve o hay nubes, pierdes el dinero y el tiempo. Yo siempre recomiendo dejar al menos dos días posibles en el itinerario y decidir según la previsión meteorológica. <strong className="text-foreground">Segundo error: ir en pleno verano esperando la postal nevada.</strong> De junio a septiembre la nieve de la cumbre desaparece casi por completo, y la montaña se ve marrón y pelada. Bonita, pero no es la imagen que tienes en la cabeza. <strong className="text-foreground">Tercer error: no comprobar la visibilidad antes de salir.</strong> Hay webcams en tiempo real del Fuji (busca "Fujisan live camera") que te muestran exactamente lo que vas a encontrar. Revísalas antes de comprar tu billete.
-            </p>
-
-            <h3 className="text-xl font-medium text-foreground mt-8 mb-4">
-              Consejos de un guía local
-            </h3>
-            <p className="text-muted-foreground leading-relaxed mb-4">
-              Más allá de lo básico, estos son consejos que solo aprendes con experiencia sobre el terreno. <strong className="text-foreground">Lleva monedas de 100 yenes</strong>: los baños en la zona del Fuji y en la ruta de escalada cuestan 200-300 yenes y no aceptan tarjeta. Compra tu almuerzo en un konbini (tienda de conveniencia) en Tokio antes de salir: los onigiri y sándwiches cuestan la mitad que la comida en la zona turística de Kawaguchiko. Si vas a la zona de los Cinco Lagos, alquila bicicleta en Kawaguchiko en lugar de depender del autobús local, que pasa con poca frecuencia. Y si viajas con niños o personas mayores, la Quinta Estación del Fuji (accesible en autobús) ya ofrece una experiencia memorable con vistas impresionantes, tiendas de souvenirs y un pequeño santuario sintoísta, sin necesidad de escalar ni un metro.
-            </p>
-
-            {/* Preguntas frecuentes */}
-            <h2 className="heading-section text-foreground mt-12 mb-6">
-              Preguntas frecuentes sobre el Monte Fuji desde Tokio
-            </h2>
-
-            <div className="space-y-6 mb-8">
-              <div>
-                <h3 className="text-xl font-medium text-foreground mb-2">
-                  ¿Cuánto cuesta ir al Monte Fuji desde Tokio?
-                </h3>
-                <p className="text-muted-foreground leading-relaxed">
-                  El transporte ida y vuelta cuesta entre 4.400 yenes (autobús) y 8.260 yenes (tren expreso). A eso hay que sumar el autobús local hasta la Quinta Estación si quieres subir (unos 2.100 yenes ida y vuelta) y la tasa de acceso a la montaña (2.000 yenes). Un día completo con comida y transporte sale por unos 8.000-15.000 yenes por persona, dependiendo de las opciones que elijas.
-                </p>
-              </div>
-
-              <div>
-                <h3 className="text-xl font-medium text-foreground mb-2">
-                  ¿Cuántas horas se tarda?
-                </h3>
-                <p className="text-muted-foreground leading-relaxed">
-                  Llegar a la zona de Kawaguchiko toma entre 1 hora 50 minutos (tren) y 2 horas (autobús) desde Shinjuku. Si solo quieres ver el Fuji y pasear por la zona, 4-5 horas en destino son suficientes. Si vas a escalar, necesitas un mínimo de 8-10 horas para la subida y bajada desde la Quinta Estación.
-                </p>
-              </div>
-
-              <div>
-                <h3 className="text-xl font-medium text-foreground mb-2">
-                  ¿Se puede ir y volver en un día?
-                </h3>
-                <p className="text-muted-foreground leading-relaxed">
-                  Sí, perfectamente. La mayoría de mis clientes hacen la excursión al Monte Fuji desde Tokio en un día, saliendo temprano por la mañana y volviendo por la tarde-noche. Incluso la escalada se puede hacer en un día si estás en buena forma y sales antes del amanecer, aunque yo recomiendo pernoctar en un refugio para ver el amanecer desde la cumbre.
-                </p>
-              </div>
-
-              <div>
-                <h3 className="text-xl font-medium text-foreground mb-2">
-                  ¿Se ve el Monte Fuji todo el año?
-                </h3>
-                <p className="text-muted-foreground leading-relaxed">
-                  La montaña está ahí todo el año, pero la visibilidad varía enormemente. En verano, las nubes y la humedad lo ocultan la mayoría de los días. En invierno (diciembre-febrero), la probabilidad de verlo con nitidez es mucho mayor. Como guía, yo siempre digo que en invierno lo ves 3-4 días de cada 5; en verano, quizás 1 de cada 5.
-                </p>
-              </div>
-
-              <div>
-                <h3 className="text-xl font-medium text-foreground mb-2">
-                  ¿Necesito guía para subir al Monte Fuji?
-                </h3>
-                <p className="text-muted-foreground leading-relaxed">
-                  No es obligatorio. La Ruta Yoshida está bien señalizada y es transitada por miles de personas cada temporada. Sin embargo, un guía puede marcar la diferencia en la gestión del ritmo, la aclimatación a la altitud, la elección del mejor momento para iniciar la caminata y la resolución de imprevistos. Para la visita sin escalada a la zona de los Cinco Lagos, un guía local te ayuda a optimizar el tiempo y descubrir lugares que las guías no mencionan.
-                </p>
-              </div>
-            </div>
-
-            {/* CTA */}
-            <div className="bg-secondary/50 rounded-lg p-8 mt-12">
-              <h2 className="text-2xl font-medium text-foreground mb-4">
-                ¿Quieres visitar el Monte Fuji desde Tokio con un guía local?
-              </h2>
-              <p className="text-muted-foreground leading-relaxed mb-6">
-                Organizo excursiones privadas al Monte Fuji adaptadas al clima del día, tus intereses y tu ritmo. Nada de autobuses turísticos con 40 personas ni paradas en tiendas de souvenirs obligatorias. Solo tú, la montaña y un guía que conoce cada rincón de la zona. Si prefieres, también puedo integrar la visita al Fuji dentro de un itinerario más amplio por Tokio y alrededores.
-              </p>
-              <div className="flex flex-col sm:flex-row gap-4">
-                <Link to="/es/tours" className="btn-accent">
-                  Ver Nuestros Tours
-                </Link>
-                <Link to="/es/contact" className="btn-outline">
-                  Hacer una Pregunta
-                </Link>
-              </div>
-            </div>
-            <ArticleFooterCTAEs />
-          </article>
-        </div>
-      </section>
-
-      <RelatedTourCards tourIds={["custom"]} lang="es" showViewAll />
-
-      {/* BlogPosting Schema */}
-      <script
-        type="application/ld+json"
-        dangerouslySetInnerHTML={{
-          __html: JSON.stringify({
-            "@context": "https://schema.org",
-            "@type": "BlogPosting",
-            "headline": "Monte Fuji desde Tokio: Guía Completa 2026 para Verlo o Subir",
-            "description": "¿Cómo ir al Monte Fuji desde Tokio? Un guía con licencia oficial explica las mejores opciones para verlo o subirlo, temporadas, precios y errores que evitar.",
-            "author": {
-              "@type": "Person",
-              "name": "Manabu",
-            },
-            "datePublished": "2026-03-07",
-            "publisher": {
-              "@type": "Organization",
-              "name": "Tanuki Tabi Travel",
-              "url": "https://tanuki-tabi-travel.com",
-            },
-            "mainEntityOfPage": {
-              "@type": "WebPage",
-              "@id": "https://tanuki-tabi-travel.com/es/blog/monte-fuji-desde-tokio",
-            },
-          }),
-        }}
-      />
-    </Layout>
-  );
-};
-
-export default EsMonteFujiDesdeTokio;
